biopass_dao/requirements.txt
psycopg2-binary
python-dotenv
opencv-contrib-python
numpy
